Temporary patches to fix building with VS2017.

+++ lib/jxl/modular/encoding/ma.cc
@@ -576,7 +576,7 @@ Status TreeSamples::SetProperties(const std::vector<uint32_t> &properties,
                                   ModularOptions::WPTreeMode wp_tree_mode) {
   props_to_use = properties;
   if (wp_tree_mode == ModularOptions::WPTreeMode::kWPOnly) {
-    props_to_use = {kWPProp};
+    props_to_use = {static_cast<uint32_t>(kWPProp)};
   }
   if (wp_tree_mode == ModularOptions::WPTreeMode::kNoWP) {
     auto it = std::find(props_to_use.begin(), props_to_use.end(), kWPProp);
